==> analysis/00.reddit_preprocessing.ipynb <==
import pandas as pd
import re

# Load the CSV files
df1 = pd.read_csv('data/comment.csv')
df2 = pd.read_csv('data/submission.csv', encoding='ISO-8859-1')

# Extract subreddit from 'link' column using regex
def extract_subreddit(link):
    match = re.search(r'reddit\.com/r/([^/]+)/', str(link))
    return match.group(1) if match else None

df1['subreddit'] = df1['link'].apply(extract_subreddit)
df2['subreddit'] = df2['link'].apply(extract_subreddit)

# Add type only to df1
df1['type'] = 'comment'

#'created' to 'created_utc' 
df1.rename(columns={'created': 'created_utc'}, inplace=True)
df2.rename(columns={'created': 'created_utc'}, inplace=True)

#  Select and align columns (now including 'score')
df1_final = df1[['author', 'subreddit', 'body', 'created_utc', 'score', 'type']]
df2_final = df2[['author', 'subreddit', 'body', 'created_utc', 'score', 'type']]

# Merge
merged_df = pd.concat([df1_final, df2_final], ignore_index=True)

merged_df.head()

merged_df.to_csv('data/20250706_reddit_merged.csv', index=True)

df3 =merged_df

df3.shape

# Convert the 'created_utc' column to datetime
df3['created_utc'] = pd.to_datetime(df3['created_utc'])

# Extract the year from the 'created_utc' datetime object and create a new column
df3['year'] = df3['created_utc'].dt.year

import matplotlib.pyplot as plt
plt.style.use('default')

# Plot a histogram
plt.bar(yearly_counts.index, yearly_counts.values)
plt.xlabel('Year')
plt.ylabel('Number of Posts')
plt.title('Yearly Count of Posts')
plt.xticks(yearly_counts.index, rotation=45)  
plt.tight_layout()  
plt.show()

#duplication removal
df3 = df3.drop_duplicates(subset=['body'])
df3.shape
#no rows are removed


#null value removal. 
df3 = df3.dropna(subset=['created_utc','body'])
df3.shape
# 2,434 rows are removed

# Replace '[deleted]' or 'NA' with an empty string
df3.replace({'\[deleted\]': '', '\\bNA\\b': ''}, regex=True, inplace=True)

df3.shape

#Exlcuding if the body contains strings less than 5 characters
df3 = df3[df3['body'].str.len() >= 5]

df3.shape
#18 rows are romoved

#converting all body of text to lower cases
df3['body'] = df3['body'].str.lower()

df3.head()

#prep done
df3.to_csv("results/reddit_prep.csv",index=False)
